+++ src/events/init.ts
@@ -1,0 +1,3 @@
+export class InitStore {
+  static readonly type = '@@INIT';
+}