+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+# 1.1.0 2/3/18
+- Fix: Export plugin interface
+
+# 1.1.0 2/3/18
+- Feature: Plugins improvements
+- Feature: Init event
+- Feature: Logger plugin
+
 # 1.0.4 2/2/18
 - Inital release!